pyproject.toml
[tool.poetry]
name = "wyn-voice"
version = "0.1.9"
description = "This library provides audio interface with OpenAI endpoint."
authors = ["Yiqiao Yin <eagle0504@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.35.14"
ipython = "7.34.0"
google-colab = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
